Set up for trait canister api
Problem: Since we want to treat factory canisters not as a
separate `extend_with_factory_api!()` macro that extends the api
of the given canister, but treat the given canister as a subtrait
of `FactoryCanister` trait, we need to do some refactorings on
deriving macros side. More specifically, we have to make the
internal generated async methods synchronous and return a pinned
box of the future instead. This posseses a problem with idl
generation since `CandidType` is not implemented for
`Pin<Box<impl Future<Output = T>>>`, but we can come up with a
little hack. Since this future is called correctly inside ic_cdk
call, we can extract the type of the future when storing it in
`METHODS` lazy static, that will be later queried when generating
the idl for the canister.