Set up factory canister trait
We want to reduce the amount of macros, used by our ic canisters.
One of them is `extend_with_factory_api!()`, which extends the
given canister with factory queries. We agreed that we'll take a
tagless final approach and make a `FactoryCanister` trait, that
will be a supertrait over all the canisters that user wants to be
extended with factory api.