Add export_methods!() macro
We have decided to change api a bit, because, unfortunately, rust
traits do not allow for attributes in their methods. That is
because monomorphisation in rust is optional to allow trait objects
meaning that while compiling to wasm, there's no way to know what
actual method will be used when dynamically dispatched. So, now
what we've came up with is that for the purpose of exporting
methods to wasm, we need to create an arbitrary struct, trait of
the canister of which will be implemented for and call,
`export_methods!()` on that struct. For an example you can refer
to the `canister_a` crate.