Add attribute to allow custom upgrade methods in ic-canister
This PR changes the way the canister state is instantiated in test context. Instead of keeping the state for each canister instance in a structure generated by `Canister` derive macro, we move this HashMap to the implementation of ic-storage macros.

This way the state can safely be used as is in traits and structs alike. E.g. every call to `IcStorage::get()` will always consider test context and produce correct result both in test environment and in wasm canister.

Using this approach, we can now easily store the factory state inside the trait, exposing it by the factory trait method.

With this approach implementing custom upgrade methods for canisters using multiple state structures becomes much easier. To allow this we introduce `#[canister_no_upgrade_methods]` attribute that instructs the derive macro not to generate the upgrade methods.