Fixes tests context ids
It turns out that changing the text context ids on simple calls produces some undesirable side effects. For example, when a canister calls its own methods, with such approach the caller id and current canister id get messed up, which breaks many tests.

Instead, we rely on the test author to provide the test context when calling the canister methods directly, and make the `canister_call!` macros substitute the context. This approach is more transparent and less error prone.