Fix factory upgrade method
This commit fixes two issues:
1. The upgrade method in the factory API used `FactoryCanister::get_canister_bytecode` method. When such method is accessed from the FactoryExport struct (the one that is created just for trait APIs), which doesn't and cannot provide an implementation for this method. Instead, we remove this API method from the trait, leaving the actual method logic in the trait. This way, the factory structure implementing the `FactoryCanister` trait can have a one-line API method implementation that just calls the method in the trait providing the wasm code as an argument.
2. The `pre_upgrade` and `post_upgrade` macros reuse the code for other types of API methods, which included requesting the method call arguments from IC. Such a call resulted in a runtime error in a canister when trying to upgrade. We remove the arguments request for `pre_` and `post_upgrade` methods, and add compile-time check that those methods don't expect such arguments.

Due to the discovery of the first issue, it's now clear that a canister trait actually cannot safely have unimplemented methods (because of how the API methods are exported). This means that the export internal structure will always be an empty impl block, so we can move this struct generation into the `generate_exports` macro, which simplifies the usage of canister traits and prevent errors when writing export structs.