[CPROD-1056] Remove default implementation of state getters in traits
Problem: As there's a default `State::get()` call that is used
in a trait canister there's a problem that it's only crate-local.
Hence, the user has to remember to overwrite it or he'll be forced
to face the UB as if there's no stable state yet it'll return
a default state. We should add a macro that has compile-time checks
for state getters and that helps to add a default implementation
of state getter inside `generate_exports!` struct (we don't really
use the state there, only the interface export).